pyproject.toml
[tool.poetry]
name = "document-intelligence-backend"
version = "0.1.0"
description = "Async document processing backend"
authors = ["Kolade"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110"
uvicorn = "^0.29"
pydantic-settings = "^2.2"
python-json-logger = "^4.0.0"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
alembic = "^1.17.2"
psycopg2-binary = "2.9.11"
python-jose = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.21"
minio = "^7.2.20"
boto3 = "^1.42.16"
redis = "<6.0.0"
celery = {version = "5.3.4", extras = ["redis"]}
pytesseract = "^0.3.13"
pillow = "^12.0.0"
pdf2image = "^1.17.0"
python-docx = "^1.2.0"
pypdf2 = "^3.0.1"
pdfminer-six = "^20251228"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
google-generativeai = "^0.8.6"
google-genai = "^1.56.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"